third_party/WebKit/LayoutTests/fast/workers/shared-worker-shared-expected.txt

Test simple shared worker sharing cases. Should print several PASS lines followed by DONE.

PASS: Exception thrown when creating SharedWorker with different URLs but same name: URLMismatchError: A worker global scope represented an absolute URL that is not equal to the resulting absolute URL.
PASS: Accessing new instance of shared worker: self.foo: undefined
PASS: Setting global variable in shared worker: self.foo = 1234: 1234
PASS: Accessing simultaneously-loaded instance of shared worker: self.foo: 1234
PASS: Accessing new instance of shared worker: self.foo: undefined
PASS: Accessing already-loaded instance of shared worker: self.foo: 1234
DONE
